requirements.txt
opencv-python>=4.5
numpy>=1.20
flask
picamera[array]
toml
gphoto2
pyserial
gpiozero
RPi.GPIO
smbus2
vl53l1x
# imutils
